requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
albumentations==0.4.6
astroid @ file:///tmp/build/80754af9/astroid_1592495881661/work
async-timeout==3.0.1
attrs==20.3.0
blessings==1.7
bottleneck-transformer-pytorch==0.1.2
cachetools==4.2.1
catalyst==20.12
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1596809855670/work
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Colr==0.9.1
configparser==5.0.2
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1595348769214/work
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
docker-pycreds==0.4.0
docstring-parser==0.7.3
efficientnet-pytorch==0.6.3
einops==0.3.0
fire==0.4.0
fsspec==0.9.0
future==0.18.2
fvcore @ git+https://github.com/facebookresearch/fvcore.git@bec6c43b3cf4a2236a480a1c79ed02b9901b9907
gcc7==0.0.7
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
gpustat==0.6.0
graphviz==0.16
grpcio==1.37.0
grpcio-tools==1.37.0
helpdev==0.7.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.4.0
imutils==0.5.4
iniconfig==1.1.1
iopath==0.1.6
isort==4.3.21
joblib==1.0.1
kaggle==1.5.12
kaptan==0.5.12
kiwisolver==1.3.1
labml==0.4.112
lazy-object-proxy==1.4.3
libtmux==0.8.5
Markdown==3.3.4
matplotlib==3.3.4
mccabe==0.6.1
minio==7.0.3
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mpi4py==3.0.3
mpmath==1.2.1
multidict==5.1.0
multiprocess==0.70.11.1
munch==2.5.0
networkx==2.4
nibabel==3.2.1
ninja==1.10.0.post2
numpy==1.20.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.40
opencv-python-headless==4.5.1.48
p-tqdm==1.3.3
packaging==20.9
pandas==1.2.2
pathos==0.2.7
pathtools==0.1.2
Pillow @ file:///tmp/build/80754af9/pillow_1594307325547/work
pluggy==0.13.1
portalocker==2.2.1
pox==0.2.9
ppft==1.6.6.3
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.14.0
psutil==5.8.0
py==1.10.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pylint @ file:///tmp/build/80754af9/pylint_1592496044040/work
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.1
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-slugify @ file:///home/conda/feedstock_root/build_artifacts/python-slugify_1593573453419/work
pytorch-gradcam==0.2.1
pytorch-lightning==1.2.8
pytorchcv==0.0.58
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
QDarkStyle==2.8.1
QtPy==1.9.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
-e git+git@gitlab.com:ficsresearch/s3a.git@c4e49a916e7cfeea7f141f3d3e35d2dd7bbfd9f4#egg=s3a
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
sentry-sdk==1.0.0
Shapely==1.7.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
sympy==1.8
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.8.13
tmuxp==1.7.2
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
torch==1.8.1
torch-tb-profiler==0.1.0
torchaudio==0.7.0a0+a853dff
torchmetrics==0.2.0
torchsummary==1.5.1
torchvision==0.9.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1596476591553/work
triton==1.0.0.dev20210330
typed-ast==1.4.1
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
Unidecode==1.1.1
urllib3==1.24.3
utilitys==0.2.0
wandb==0.10.30
Werkzeug==1.0.1
wrapt==1.11.2
xmltodict==0.12.0
-e git+git@gitlab.com:ficsresearch/xrayrecon.git@8e144698091ded718c55e5df4c03b743079d0e82#egg=xrayrecon
yacs==0.1.8
yarl==1.6.3
zipp==3.4.0
